Avance 26/06/2020

+++ app/src/main/java/com/fuenteoro/ciclismo/Models/Rutas.java
@@ -2,12 +2,12 @@ package com.fuenteoro.ciclismo.Models;
 
 public class Rutas {
     private String nombre, distancia, elevacion, dificultad;
-    private Long latitud_origen, longitud_origen;
-    private Long latitud_destino, longitud_destino;
+    private Double latitud_origen, longitud_origen;
+    private Double latitud_destino, longitud_destino;
     private String imagen;
     private int calificacion;
 
-    public Rutas(String nombre, String distancia, String elevacion, String dificultad, Long latitud_origen, Long longitud_origen, Long latitud_destino, Long longitud_destino, String imagen, int calificacion) {
+    public Rutas(String nombre, String distancia, String elevacion, String dificultad, Double latitud_origen, Double longitud_origen, Double latitud_destino, Double longitud_destino, String imagen, int calificacion) {
         this.nombre = nombre;
         this.distancia = distancia;
         this.elevacion = elevacion;
@@ -52,35 +52,35 @@ public class Rutas {
         this.dificultad = dificultad;
     }
 
-    public Long getLatitud_origen() {
+    public Double getLatitud_origen() {
         return latitud_origen;
     }
 
-    public void setLatitud_origen(Long latitud_origen) {
+    public void setLatitud_origen(Double latitud_origen) {
         this.latitud_origen = latitud_origen;
     }
 
-    public Long getLongitud_origen() {
+    public Double getLongitud_origen() {
         return longitud_origen;
     }
 
-    public void setLongitud_origen(Long longitud_origen) {
+    public void setLongitud_origen(Double longitud_origen) {
         this.longitud_origen = longitud_origen;
     }
 
-    public Long getLatitud_destino() {
+    public Double getLatitud_destino() {
         return latitud_destino;
     }
 
-    public void setLatitud_destino(Long latitud_destino) {
+    public void setLatitud_destino(Double latitud_destino) {
         this.latitud_destino = latitud_destino;
     }
 
-    public Long getLongitud_destino() {
+    public Double getLongitud_destino() {
         return longitud_destino;
     }
 
-    public void setLongitud_destino(Long longitud_destino) {
+    public void setLongitud_destino(Double longitud_destino) {
         this.longitud_destino = longitud_destino;
     }
