+++ app/src/main/java/com/fuenteoro/ciclismo/Models/Sitios.java
@@ -2,7 +2,7 @@ package com.fuenteoro.ciclismo.Models;
 
 public class Sitios {
     private String nombre, detalle;
-    private Long latitud_sitio, longitud_sitio;
+    private Double latitud_sitio, longitud_sitio;
     private String imagen;
     private int calificacion;
 
@@ -10,7 +10,7 @@ public class Sitios {
 
     }
 
-    public Sitios(String nombre, String detalle, Long latitud_sitio, Long longitud_sitio, String imagen, int calificacion) {
+    public Sitios(String nombre, String detalle, Double latitud_sitio, Double longitud_sitio, String imagen, int calificacion) {
         this.nombre = nombre;
         this.detalle = detalle;
         this.latitud_sitio = latitud_sitio;
@@ -35,19 +35,19 @@ public class Sitios {
         this.detalle = detalle;
     }
 
-    public Long getLatitud_sitio() {
+    public Double getLatitud_sitio() {
         return latitud_sitio;
     }
 
-    public void setLatitud_sitio(Long latitud_sitio) {
+    public void setLatitud_sitio(Double latitud_sitio) {
         this.latitud_sitio = latitud_sitio;
     }
 
-    public Long getLongitud_sitio() {
+    public Double getLongitud_sitio() {
         return longitud_sitio;
     }
 
-    public void setLongitud_sitio(Long longitud_sitio) {
+    public void setLongitud_sitio(Double longitud_sitio) {
         this.longitud_sitio = longitud_sitio;
     }
